+++ src/types/Ecosystem.type.ts
@@ -3,6 +3,7 @@ interface Ecosystem {
     organization_name: string,
     responses: number,
     admin_id: string,
+    time_window: number
 }
 
 export type { Ecosystem };
